Splitting the acquisition file into a package to avoid further mixing of framework code and algorithm implementations. Paths remain consistent
