+++ GPflowOpt/acquisition/__init__.py
@@ -1,0 +1,26 @@
+# Copyright 2017 Joachim van der Herten
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+# Framework components and interfaces
+from .acquisition import Acquisition, AcquisitionAggregation, AcquisitionProduct, AcquisitionSum, MCMCAcquistion
+
+# Single objective
+from .ei import ExpectedImprovement
+from .poi import ProbabilityOfImprovement
+from .lcb import LowerConfidenceBound
+
+# Multi-objective
+
+# Black-box constraints
+from .pof import ProbabilityOfFeasibility
